% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectdata_function.R
\name{collectdata}
\alias{collectdata}
\title{collectiondata}
\usage{
collectdata(tree, assign, maj_sublineage)
}
\arguments{
\item{tree}{The rooted phylogenetic tree.}

\item{assign}{Vector of Sub-group numbers corresponding to tips on a tree.}

\item{maj_sublineage}{The output of the Major Sub-Lineage function

collectiondata()}
}
\description{
Collating data for the output.
}
