% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path_distance_function.R
\name{get_path_distance}
\alias{get_path_distance}
\title{Get the distance of a path}
\usage{
get_path_distance(graph_tree, root, node, input_tree)
}
\arguments{
\item{graph_tree}{A rooted tree of newick format, converted into graph form by igraph.}

\item{root}{A node in the graph you want to begin the path.}

\item{node}{target node.}

\item{input_tree}{a rooted tree that has been parsed into this script - class "phylo".}
}
\value{
The SNP distance between given root and node
get_path_distance()
}
\description{
This function returns the path distance between node and tip on a rooted tree, in a root to tip direction.
}
