% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itol_label_change.R
\name{itol_labels_template}
\alias{itol_labels_template}
\title{Label change Itol Output for Sub-Groups.}
\usage{
itol_labels_template(input)
}
\arguments{
\item{input}{output from pinecone}
}
\description{
Function to output the results of pinecone to itol
This will change your tip labels to the sub-group number of the isolate
}
