% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itol_major_SB_output.R
\name{itol_major_SB_output}
\alias{itol_major_SB_output}
\title{Colourstrip Itol Output Major sub-groups.}
\usage{
itol_major_SB_output(input)
}
\arguments{
\item{input}{output from maria}
}
\description{
Function to output the results of maria to itol
This will change your tip labels to the sub-group number of the isolate
}
