% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itol_major_SB_output.R
\name{itol_major_SL_output}
\alias{itol_major_SL_output}
\title{Colourstrip Itol Output Major sub-lineages.}
\usage{
itol_major_SL_output(input)
}
\arguments{
\item{input}{output from pinecone}
}
\description{
Function to output the results of pinecone to itol
}
