% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itol_subGroup_output.R
\name{itol_subGroup_output}
\alias{itol_subGroup_output}
\title{itol_subGroup_output}
\usage{
itol_subGroup_output(input)
}
\arguments{
\item{input}{Output list from main maria function.}
}
\description{
Function to output the results of maria to itol.
This will change your tip labels to the sub-group number of the isolate.
}
