% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itol_subGroup_output.R
\name{itol_sublineage_output}
\alias{itol_sublineage_output}
\title{Colourstrip Itol Output of Sub-lineage.}
\usage{
itol_sublineage_output(input)
}
\arguments{
\item{input}{Output list from main pinecone function.}
}
\description{
Function to output the results of pinecone to itol.
}
