% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tree_function.R
\name{load_tree}
\alias{load_tree}
\title{Load a Newick tree}
\usage{
load_tree(infile)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
This function loads a phylogenetic tree. The tree must be rooted, newick
format and an ACCTRANS tree.
}
