% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majorlineage_function.R
\name{majorlineage}
\alias{majorlineage}
\title{Major lineage}
\usage{
majorlineage(slnum, tree, assign, rthreshold)
}
\arguments{
\item{slnum}{Number of Sub-lineage identified.}

\item{tree}{The rooted phylogenetic tree.}

\item{assign}{Vector of lineage numbers corresponding to tips on a tree.}

\item{rthreshold}{Number of internal nodes to the root each lineage must have
intersecting to be declared as a major sub-lineage.}

\item{ntips}{Number of tips on the phylogenetic tree.}

\item{treeNnodes}{Number of internal nodes on phylogenetic tree.

majorlineage()}
}
\description{
Function to identify major lineages composed of identified lineages
Singletons which may not have a lineage can also be inlcuded in
major lineage
}
