% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majorsub_function.R
\name{majorsub}
\alias{majorsub}
\title{Major Subs}
\usage{
majorsub(sgnum, tree, assign, rthreshold)
}
\arguments{
\item{sgnum}{Number of Sub-groups identified.}

\item{tree}{The rooted phylogenetic tree.}

\item{assign}{Vector of Sub-group numbers corresponding to tips on a tree.}

\item{rthreshold}{Number of internal nodes to the root each sub-group must have
intersecting to be declared as a major sub-group.}

\item{ntips}{Number of tips on the phylogenetic tree.}

\item{treeNnodes}{Number of internal nodes on phylogenetic tree.

majorsub()}
}
\description{
Function to identify major sub-groups composed of identified sub-groups
Singletons which may not have a sub-group can also be inlcuded in
major sub-group
}
