% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maria_060218.R
\name{maria}
\alias{maria}
\title{maria}
\usage{
maria(tree, thresh, rthreshold, sample_id_file)
}
\arguments{
\item{tree}{Newick formatted ACCTRANS tree.}

\item{thresh}{SNP threshold for sub-grouping isolates.}

\item{rthreshold}{The threhold number of ancestral nodes to the root
each sub-group must have if they are to be declared as major sub-groups}

\item{sample_id_file}{CSV file, with a list of sample ids in the first column
and sample ids in the second column}
}
\description{
maria
}
