% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_tips_max_dist.R
\name{node_tips_max_dist}
\alias{node_tips_max_dist}
\title{Identify the maximum distance from an internal node amongst all
tips}
\usage{
node_tips_max_dist(node, subtree, igraph.tree, tree)
}
\arguments{
\item{node}{that is the root of the subtree}

\item{subtree}{defined by main script}

\item{igraph.tree}{a tree converted into graph format by igraph}

\item{tree}{Newick formatted ACCTRANS tree
node_tips_max_dis()}
}
\description{
Identify the maximum distance from an internal node amongst all
tips
}
