% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinecone.R
\name{pinecone}
\alias{pinecone}
\title{pinecone}
\usage{
pinecone(tree, thresh, rthreshold, quiet = FALSE)
}
\arguments{
\item{tree}{a phylo object}

\item{thresh}{SNP threshold for sub-lineage identification.}

\item{rthreshold}{The threhold number of ancestral nodes to the root
each sub-lineage must have if they are to be declared as major sub-lineages}

\item{quiet}{whether or not to print out extra information (default=FALSE)}
}
\description{
Sub-lineage identification using a phylogentic tree with branch lengths indicating SNP distance.
}
\examples{
tree.file.name <- system.file("extdata", "pyjar.staph_ST2371_45_re_itol_7079_1_6_root.joint.tre", package = "rPinecone")
tree <- phytools::read.newick(tree.file.name)
pinecone(tree, 2, 3)

}
