% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgrp_function.R
\name{subgrp}
\alias{subgrp}
\title{Subgrp}
\usage{
subgrp(dfs, tree, assign, igraph.tree, thresh)
}
\arguments{
\item{dfs}{The depth-first search performed on the igraph tree.}

\item{tree}{The rooted phylogenetic tree.}

\item{assign}{Vector of Sub-group numbers corresponding to tips on a tree.}

\item{igraph.tree}{Phylogenetic tree converted into a graph}

\item{thresh}{SNP threshold for sub-group isolates on the phylogentic tree.

subgrp()}

\item{ntips}{Number of tips on the phylogenetic tree.}
}
\description{
Function to assign sub-groups to isolates according to the structure of a rooted tree
}
