% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sublineage_function.R
\name{sublineage}
\alias{sublineage}
\title{Sub-lineage isolates on a phyologentic tree.}
\usage{
sublineage(dfs, tree, assign, igraph.tree, thresh, quiet)
}
\arguments{
\item{dfs}{The depth-first search performed on the igraph tree.}

\item{tree}{The rooted phylogenetic tree.}

\item{assign}{Vector of Sub-lineage numbers corresponding to tips on a tree.}

\item{igraph.tree}{Phylogenetic tree converted into a graph}

\item{thresh}{SNP threshold for sub-lineage isolates on the phylogentic tree.}

\item{ntips}{Number of tips on the phylogenetic tree.}
}
\description{
Function to assign sub-lineages to isolates according to the
structure of a rooted tree.
}
