% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maria.R
\name{maria}
\alias{maria}
\title{maria}
\usage{
maria(tree, thresh, rthreshold)
}
\arguments{
\item{tree}{Newick formatted ACCTRANS tree.}

\item{thresh}{SNP threshold for sub-grouping isolates.}

\item{rthreshold}{The threhold number of ancestral nodes to the root
each sub-group must have if they are to be declared as major sub-groups}
}
\description{
Grouping isolates together using an ACCTRANS tree, from root-to-tip.
}
